Get user data from SIP2, store it in DB and use that later
When a new user logs in, authenticate against SIP2 and get data about
the user from SIP2. Store the data in the database and re-set the
realm of the user to be the local database, so that the user appears
to be a local user.

If a user who has logged in before logs in again, that user is also
authenticated against SIP2, data is fetched from SIP2 and the data
in the database is *updated*. This means that if a user needs to
change e.g. their email, they will have to do it in their local ILS.

TODO
- Connect the user to the right library in the database
- Redirect users to some special info the very first time theu sign
  in, to tell them about the data we have saved and that they have
  to change it in their local ILS.