Force users to choose a library
The only thing they can do now is choose a library or log in. When
a user logs in, a library is chosen based on the realm they
authenticated against, even if they have already chosen another
library. So far the code assumes users are connected to one library,
we will fix this if/when the national library card is implemented.