Make it possible to include links to social networks on detail view
This adds two new columns: libraries.detail_head and .soc_links
The first will be included in the head of pages, the second will
be included somewhere in the detail view. (All the elements of the
detail view will probably be rearranged by a desginer later on.)

Content for these two columns can be had from e.g.
http://sharethis.com/, which will make it possible to choose from
a range of social networks as well as styles. Links can also be
added in the form of simple HTML elements.