Improve handling of large ePub files
The files.file column was not given special treatment, probably
resulting in large BLOBs being dragged around all over the place
and resulting in massive memory usage. This patches tries to
make it so that you have to request the contents of files.file
specifically to get hold of it. This should improve memory use.