Improve checking of concurrent loans
We were checking only for equality:

if ( $user->number_of_loans_from_library( $library->id ) == $library->concurrent_loans ) {

This means that if someone got more loans than the limit, they could
borrow an unlimited amount of books. This could happen if they had the max
amount of loans for their library, and the the number of concurrent loans
for their library was lowered.

Hopefully affects #25.